requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
gdown==5.2.0
requests==2.32.3
numpy==2.2.1
torchvision==0.20.1
matplotlib==3.10.0
lightning==2.5.0.post0
hydra-core==1.3.2
transformers==4.48.0
wandb==0.19.4
python-dotenv==1.0.1
httpx==0.28.1
locust==2.32.6
prometheus_client==0.21.1
google-cloud-storage==2.19.0
